+++ src/Classifiers/KDNeighbors.php
@@ -11,22 +11,21 @@ use Rubix\ML\Datasets\DataFrame;
 use Rubix\ML\Graph\Trees\KDTree;
 use Rubix\ML\Other\Functions\Argmax;
 use Rubix\ML\Kernels\Distance\Distance;
-use Rubix\ML\Kernels\Distance\Euclidean;
 use InvalidArgumentException;
 use RuntimeException;
 
 /**
  * K-d Neighbors
  *
- * A fast K Nearest Neighbors approximating algorithm that uses a K-d tree to
- * divide the training set into neighborhoods whose max size are constrained by
- * the neighborhood hyperparameter. K-d Neighbors does a binary search to locate
- * the nearest neighborhood and then searches only the points in the neighborhood
- * for the nearest k to make a prediction. Since there may be points in other
- * neighborhoods that may be closer, the KNN search is said to be approximate.
- * The main advantage K-d Neighbors has over regular KNN is that it is much
- * faster.
+ * A fast K Nearest Neighbors algorithm that uses a K-d tree to divide the training
+ * set into neighborhoods whose max size are constrained by the neighborhood
+ * hyperparameter. K-d Neighbors does a binary search to locate the nearest
+ * neighborhood and then searches only the points close to or within the neighborhood
+ * for the nearest k to make a prediction. The main advantage K-d Neighbors has over
+ * regular brute force KNN is that it is faster, however it cannot be partially
+ * trained.
  *
+ * References:
  * [1] J. L. Bentley. (1975). Multidimensional Binary Seach Trees Used for
  * Associative Searching.
  *
@@ -44,13 +43,6 @@ class KDNeighbors extends KDTree implements Learner, Probabilistic, Persistable
     protected $k;
 
     /**
-     * The distance function to use when computing the distances.
-     *
-     * @var \Rubix\ML\Kernels\Distance\Distance
-     */
-    protected $kernel;
-
-    /**
      * The unique class outcomes.
      *
      * @var array
@@ -66,7 +58,7 @@ class KDNeighbors extends KDTree implements Learner, Probabilistic, Persistable
      * @throws \InvalidArgumentException
      * @return void
      */
-    public function __construct(int $k = 3, int $neighborhood = 10, ?Distance $kernel = null)
+    public function __construct(int $k = 3, int $neighborhood = 20, ?Distance $kernel = null)
     {
         if ($k < 1) {
             throw new InvalidArgumentException("At least 1 neighbor is required"
@@ -74,19 +66,13 @@ class KDNeighbors extends KDTree implements Learner, Probabilistic, Persistable
         }
 
         if ($k > $neighborhood) {
-            throw new InvalidArgumentException("K cannot be larger than the"
-                . " size of the neighborhood, $k given but $neighborhood"
-                . " allowed.");
-        }
-
-        if (is_null($kernel)) {
-            $kernel = new Euclidean();
+            throw new InvalidArgumentException("K cannot be larger than the max"
+                . " size of a neighborhood, $k given but $neighborhood allowed.");
         }
 
         $this->k = $k;
-        $this->kernel = $kernel;
 
-        parent::__construct($neighborhood);
+        parent::__construct($neighborhood, $kernel);
     }
 
     /**
@@ -153,7 +139,7 @@ class KDNeighbors extends KDTree implements Learner, Probabilistic, Persistable
                 . ' continuous features.');
         }
 
-        if ($this->bare() === true) {
+        if ($this->bare()) {
             throw new RuntimeException('Estimator has not been trainied.');
         }
 
@@ -161,7 +147,7 @@ class KDNeighbors extends KDTree implements Learner, Probabilistic, Persistable
             array_fill_keys($this->classes, 0.));
 
         foreach ($dataset as $i => $sample) {
-            $neighbors = $this->findNearestNeighbors($sample);
+            $neighbors = $this->neighbors($sample, $this->k);
 
             $n = count($neighbors);
 
@@ -171,32 +157,5 @@ class KDNeighbors extends KDTree implements Learner, Probabilistic, Persistable
         }
 
         return $probabilities;
-    }
-
-    /**
-     * Find the k nearest neighbors to the given sample from a neighborhood.
-     *
-     * @param  array  $sample
-     * @return array
-     */
-    public function findNearestNeighbors(array $sample) : array
-    {
-        $neighborhood = $this->search($sample);
-
-        if (is_null($neighborhood)) {
-            return [];
-        }
-
-        $distances = [];
-
-        foreach ($neighborhood->samples() as $neighbor) {
-            $distances[] = $this->kernel->compute($sample, $neighbor);
-        }
-
-        asort($distances);
-
-        return array_intersect_key($neighborhood->labels(),
-            array_slice($distances, 0, $this->k, true));
-
     }
 }