+++ CHANGELOG.md
@@ -2,6 +2,7 @@
     - Added Verbose interface for logging
     - Added Linear Discriminant Analysis (LDA) transformer
     - Added Cyclic neural net Optimizer
+    - Added k-d neighbors search with pruning
     - Estimators with explicit loss functions are now Verbose
     - Grid Search: Added option to retrain best model on full dataset
     - Filesystem Persister now keeps backups of latest models